// backend/api/src/common/enums/validation-msg.enum.ts
export enum ValidationMessage {
  USERNAME_REQUIRED = 'Username tidak boleh kosong',
  USERNAME_MIN = 'Username minimal 3 karakter',
  USERNAME_MAX = 'Username maksimal 50 karakter',
  TOKEN_REQUIRED = 'Token tidak boleh kosong',
  PASSWORD_REQUIRED = 'Password tidak boleh kosong',
  PASSWORD_MIN = 'Password minimal 6 karakter',

  REFRESH_EMPTY = 'Refresh token tidak boleh kosong',
}